+++ SurvivalGame/Assets/WorldObjects/WorldObjectData.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+public class WorldObjectData : ScriptableObject
+{
+    public string objectName;
+}